+++ src/main/java/healthcalc/Adapter/HealthHospital.java
@@ -4,6 +4,6 @@ public interface HealthHospital {
 	
 	public double idealWeightA(char gender, float height) throws Exception;
 	
-	public float bmrA(char gender, int age, float height, int weight) throws Exception;
+	public double bmrA(char gender, int age, float height, int weight) throws Exception;
 
 }
